refactor(interop): plan-coverage-groups.ts -> src/core/specs.ts
Nothing in it plans groups any more — the matrix is hardcoded and the packer is
gone, leaving discoverSpecs and assertEverySpecRan. It was also the only file at
the package root that is not an executable entry point; it sat there because it
used to be the `plan:groups` CLI.

Both functions stay: assertEverySpecRan is exactly what makes a hand-written
matrix safe, since a spec added to the repo but not to a group would otherwise
never run with every job still green.

Renamed the test alongside it, dropped the header's account of a planner that no
longer exists, and rewrote SAFE_SPEC_NAME's comment to the boundary that still
holds — spec names reach a shell through the matrix.
